Part 8.2

+++ src/main/java/com/parkingvspb/igor_sasha/parking/repository/DetailsRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DetailsRepository extends JpaRepository<UserDetails, Integer> {
+public interface DetailsRepository extends JpaRepository<UserDetails, Long> {
 }
